Hack in new Immix allocator policy.
This hack introduces a heap growth policy into the allocator. Ultimately,
the policy needs to define the allocator, or the allocator should be given
an external policy to interact with. The existing Trigger thing may have
been intended to do this but doesn't provide interaction (it's one-way).

The goal right now is to stabilize heap growth so that concurrent marking
can be re-introduced, then the young gen can be re-added.