Avoid operating on (really) dead Threads.
The issue here is that Thread instances may be retained by other objects (eg
Enumerator) and when the Thread instance itself is finalized, the order of
finalization is non-deterministic (ie there is no Ruby API to impose an
ordering on finalization), so a Thread instance may have had its memory for
VM* cleaned up, and the other object referring to it may call methods on it
when that other object is being finalized.

So, we avoid operating on *really* dead Threads, where a Thread's lifetime
looks something like this:

1. Memory for the Thread allocated, but the Thread is not running as a
separate thread of execution.
2. The Thread is running.
3. The Thread's separate thread of execution has exited, so it is put in a
"zombie" state, but all the memory for the Thread (except it's execution
stack) is retained.
4. The Thread itself is still allocated (ie because the Thread is in a
finalization list so its managed memory hasn't been recycled, or it's just
been finalized) but the Thread isn't reachable by any live objects.
5. The Thread has been finalized, so it's VM* memory is deallocated, but the
managed memory for the Thread itself hasn't been reclaimed.
6. The Thread's managed memory is no longer allocated and any reference to the
Thread would be an "access after free" memory violation (but not necessarily
one that would trigger a memory fault in the CPU because the region of memory
it inhabited may still be allocated to the running program).