Eliminate extra halt lock.
Since the phase_lock is recursive, and we *always* need to acquire the lock
when transitioning to managed phase, the halting thread can just holt this
lock unconditionally while halting. That thread will then be able to run
stop-the-world garbage collection with the normal process which acquires and
releases the phase lock. This eliminates the second halt lock.