Reset logger lock on fork.
When calling fork(), we hold all running threads to keep them in known states
across the call to fork().

In the child process, the only thread that persists is the thread that called
fork(). If another thread was about to write to the logger and got suspended
before unlocking the logger mutex, the mutex will be locked in the child and
will block any other thread that tries to lock it. The thread in the parent
that would have unlocked the lock won't exist in the child.

An alternative would be for the thread calling fork to hold all the relevant
locks before calling fork(), and then unlocking them all after fork() in the
child. Functionally, that wouldn't be any different and would impose
additional costs before the fork().

All code needs to be audited for any locks that could be held across a fork()
call and all those should be re-init'd in the child.