Improved logging of subprocess creation.
The log will now reliably include the non-core method that invoked the spawn,
backtick, fork, or exec command. This will point to the application code or
rubygems code, rather than the Rubinius core library. The location in the app
code is much more useful typically than the location in the Rubinius core
library code.