Improved Fiber attributes and logging.
Fiber#source returns the non-core library source code line where the Fiber
was created. Fiber#thread_name returns the name of the Thread that created the
Fiber. Fiber#fiber_id returns a unique ID for the Fiber instance.

The Rubinius log includes entries for the Fiber create and exit events.