Preserve ThreadNexus phase_flag state.
The ThreadNexus::phase_flag is part of the thread synchronization mechanism
and additionally used to put the system into a sort of 'singled-threaded mode'
that can run arbitrary code. In order to preserve this, we have to know that
when we 'acquired' the lock whether the lock was already held by us. If it
was, we don't want to release it.