Slightly relax logger semaphore threshold.
Rather than possibly deadlocking, we reset the logger semaphore if it appears
to be in a bad state. This means it's possible that incorrect interleaving
could occur in the log file. We accept this tradeoff for not deadlocking a
process.

We attempt to make the threshold high enough to avoid spurious failures but
not high enough to lock up a process too long because a Thread may write to
the log when the process is trying to checkpoint for eg garbage collection.