Convert instructions to inlined functions.
The instructions themselves are snippets of code that mutate state.

The *interpreter* executes the functions and moves the instruction pointer
(IP) to the next instruction, essentially operating as a loop (implemented by
tail calls) over the instruction (opcode) vector.

The naive JIT *linearizes* the sequence of function calls to the instructions
and introduces control flow directly (instead of it being an artifact of
iterating over the opcode vector).