Moved builtin/ to class/.
The name 'builtin' is not very descriptive. Now that we are going to have
managed objects that are both object-oriented and simple data, and both are
'built in', we need more descriptive names for the directories, so the new
names are 'class' and 'data', which parallels the names used in the code.