Some tests for cast_for_multi_block_arg.
This instruction will call the method #to_ary on the object in the arguments
struct. This object needs to have a custom #to_ary method to test the various
logic branches in the instruction. The simplest way to do this is to create a
one-off class and attach a C++ method as a primitive implementation of the
method #to_ary.