Retire insns/interp machine tests.
Since starting on the new interpreter and instruction set, WebAssembly has
been officially adopted by all major browser vendors and has become a
legitimate implementation target.

To facilitate transitioning to multiple VM targets, the instruction set needs
to be specified in a format that is target-independent. The spec/instructions
specs are a reasonable intermediate target, but are also a step towards a
language-neutral textual asm format and assembler that will provide complete
targe independence.