Convert CodeDB to a single file.
The CodeDB cache meets three constraints:

1. easy to lock for concurrent access;
2. easy to mmap for low latency;
3. easy to update the data wihout writing a lot of data.

The single file is composed of the following regions:

1. header includes the file magic, cache version, signature, and region
   sizes;
2. list of code IDs to run when initializing the system from this cache;
3. the code itself in serialized form (data);
4. the index of code IDs to data offset and length;
5. the contents mapping files to code IDs of the file execution context.

When the cache is opened, the data region is mmap'd for low latency
access to loading code. The index and contents regions are read into
in-memory map structures.

When the cache is used in a read-write mode, on close, the new data is
appended to the existing data by writing only the new data at the end of
the data region. The index and contents maps are then written to the new
index and contents regions after the data region.