Remove use of SecureRandom.
This also assigns a random code_id to a CompiledCode object. In the
future, the code_id will be assigned based on the serialized form of the
CompiledCode object to better represent the uniqueness of a particular
instance but also be stable (with a seed) to enable using the code_id as
a key for associating other data.