Rework CallSite caching.
This restructures CallSite and the inline caching concepts slightly from
the previous implementation.

A CallSite is a location in the source code where another method is
invoked. A CallSite may or may not have a cache.

A CallSite tracks every time it is invoked. The default invocation of a
CallSite that has never been invoked does not cause the first method to
be cached. This is to avoid caching methods that will only be inovked
once.

A CallSite cache has a max size. New cache entries are always added at
the front of the list of cache entries. During a garbage collection
cycle, invalid cache entries are evicted and the list of entries is
re-ordered so that more "efficient" caches occur in the list before less
efficient ones. The efficiency is calculated as (h / (h + m) * h), where
h is hits, m is misses, and the scaling by h is intended to more
realistically represent cases like 100/100 and 2/2. The former is more
"efficient" (ie more likely to be hit) than the latter.

A CallSite that has frequently evicted cache entries is likely
megamorphic and will not benefit from caching as the cost of mutating
the cache and iterating to find a possibly matching entry outweighs just
dispatching to the method. In this case, the CallSite will disable
caching and just dispatch every time the CallSite is invoked.