Preserve C-API handle references for several GC cycles.
In the MRI C-API, there are no explicit 'retain', 'release' operations.
When a managed object is created in Rubinius, a static (ie non-moving)
pointer is returned. The pointer is a MemoryHandle object that
references the managed memory object.

When a GC cycle occurs, if nothing has marked the managed memory object
as reachable, the MemoryHeader object would be released and the managed
memory object's space reclaimed.

The problem with this is that any time the C-API calls back into the VM,
a GC cycle may occur. The following code illustrates this:

  VALUE c_func(VALUE a, VALUE b) {
    VALUE new_obj = c_api_create_obj(a, b);

    /* Many C-API functions are implemented as wrappers for Ruby code,
     * so this call may result in a GC operation and 'new_obj' will not
     * be referenced anywhere, causing the GC to assume it is
     * unreachable, and therefore garbage.
     */
    other_c_api_func(new_obj);

    return new_obj;
  }

To address this, we track how many GC cycles have occurred while the
object is not marked. To make this slighty more robust (ie not retain
true garbage for too long), we could use a flag that the MemoryHandle
has just been created and only retain it for several cycles if it has
just been created.