Re-introduce ByteArray pinning temporarily.
It appears that all operations with a pinned ByteArray (or String
containing a pinned ByteArray) need to be checked to propagate
correctly. Since we don't yet compact (evacuate) Immix regions, there's
no immediate impact, but when that is implemented, we'll need to
consider it. It may be possible to introduce a RByteArray (similar to
RTuple) to accommodate the C-API.